+++ tests/testthat/test_2_5_textPredict.R
@@ -76,30 +76,32 @@ test_that("textPredictTest t-test and bootstrapped test", {
 
 test_that("1. textPredict generates embeddings from text and 2. automatically codes implicit motives", {
   skip_on_cran()
-  
+
   # Test data
   implicit_motive_data <- dplyr::mutate(.data = text::Language_based_assessment_data_8, participant_id = dplyr::row_number(), story_id = rep(1:5, each=8))
-  
-  predictions <- textPredict(texts = implicit_motive_data$satisfactiontexts,
-                             model_info = "implicit_power_roberta_large_L23_v1",
-                             participant_id = implicit_motive_data$participant_id,
-                             story_id = implicit_motive_data$story_id,
-                             dataset_to_merge_predictions = implicit_motive_data, 
-                             previous_sentence = TRUE)
-  
+
+  predictions <- text::textPredict(
+    texts = implicit_motive_data$satisfactiontexts,
+    model_info = "implicit_power_roberta_large_L23_v1",
+    participant_id = implicit_motive_data$participant_id,
+    story_id = implicit_motive_data$story_id,
+    dataset_to_merge_predictions = implicit_motive_data,
+    previous_sentence = TRUE
+    )
+
   testthat::expect_is(predictions$sentence_predictions$texts[1], "character")
   testthat::expect_equal(predictions$person_predictions$person_prob[40], -0.957266, tolerance = 0.0001)
-  
-  # Observe; when converting to numeric, zeros are replaced by ones, and ones are replaced by twos.  
+
+  # Observe; when converting to numeric, zeros are replaced by ones, and ones are replaced by twos.
 
   # sentence predictions
   testthat::expect_equal(as.numeric(predictions$sentence_predictions$power_class[24]), 1, tolerance = 0.0001)
   testthat::expect_equal(sum(as.numeric(predictions$sentence_predictions$power_class)), 190, tolerance = 0.0001)
-  
-  # person-level predictions 
+
+  # person-level predictions
   testthat::expect_equal(sum(as.numeric(predictions$person_predictions$participant_id[10])), 10, tolerance = 0.0001)
   testthat::expect_equal(sum(as.numeric(predictions$person_predictions$person_prob[10])), -0.957266, tolerance = 0.0001)
-  
+
   # story-level predictions
   testthat::expect_equal(sum(as.numeric(predictions$story_predictions$story_id[5])), 5, tolerance = 0.0001)
   testthat::expect_equal(sum(as.numeric(predictions$story_predictions$story_prob[5])), -0.56957, tolerance = 0.0001)